+++ nav2_route/include/nav2_route/route_planner.hpp
@@ -70,8 +70,8 @@ public:
    * @return Route object containing the navigation graph route
    */
   virtual Route findRoute(
-    Graph & graph, unsigned int start_index, unsigned int goal_index,
-    const std::vector<unsigned int> & blocked_ids,
+    Graph & graph, uint64_t start_index, uint64_t goal_index,
+    const std::vector<uint64_t> & blocked_ids,
     const RouteRequest & route_request);
 
 protected:
@@ -90,7 +90,7 @@ protected:
    */
   void findShortestGraphTraversal(
     Graph & graph, const NodePtr start_node, const NodePtr goal_node,
-    const std::vector<unsigned int> & blocked_ids,
+    const std::vector<uint64_t> & blocked_ids,
     const RouteRequest & route_request);
 
   /**
@@ -101,7 +101,7 @@ protected:
    * @return if this edge is valid for search
    */
   inline bool getTraversalCost(
-    const EdgePtr edge, float & score, const std::vector<unsigned int> & blocked_ids,
+    const EdgePtr edge, float & score, const std::vector<uint64_t> & blocked_ids,
     const RouteRequest & route_request);
 
   /**
@@ -151,8 +151,8 @@ protected:
   nav2_route::EdgeType classifyEdge(const EdgePtr edge);
 
   int max_iterations_{0};
-  unsigned int start_id_{0};
-  unsigned int goal_id_{0};
+  uint64_t start_id_{0};
+  uint64_t goal_id_{0};
   NodeQueue queue_;
   std::unique_ptr<EdgeScorer> edge_scorer_;
   std::shared_ptr<tf2_ros::Buffer> tf_buffer_;
